Move IntEncodedWithFixedSize.
